Move converse-rosterview plugin into folder
