+++ src/plugins/rosterview/templates/group_header.html
