+++ src/plugins/rosterview/templates/pending_contact.html
