+++ src/plugins/rosterview/templates/requesting_contact.html
