+++ src/plugins/rosterview/templates/roster.html
