+++ src/plugins/rosterview/templates/roster_filter.js
