+++ src/plugins/rosterview/templates/roster_item.html
